+++ graphics/src/main/java/com/cgp/graphics/primitives/rasterization/matrix/ViewMatrix.java
@@ -1,4 +1,4 @@
-package com.cgp.graphics.primitives.rasterization;
+package com.cgp.graphics.primitives.rasterization.matrix;
 
 import com.cgp.graphics.entities.Camera;
 import com.cgp.graphics.entities.GameObject;
